Moving key down event handler to foreground Area View - using commands channel to communicate
